Handle functions and structs with the same name
C allows for a function and struct to have the same name. This is not
handled by ctypesgen which assignes both the function and the struct to
the same variable, effectively overriding the first one.

Modify automatically_typedef_structs() method to not alias a struct
where the corresponding function name also exists.

Signed-off-by: Ilias Tsitsimpis <iliastsi@arrikto.com>
Reviewed-by: Christos Stavrakakis <cstavr@arrikto.com>