Remove compile libdirs from output
Apart from the fact that the '_libs' variable in which the compile
directories are assigned is never used (possibly stale assignment), the
library directories used to load libraries in the compilation phase
(when the wrappers are generated) should not be exported in the final
wrappers. The final Python module should try to dynamically load the
libraries either from the default system directories or from the
'runtime-libdirs' option that the user may have specified.

Therefore, stop exporting list of the compile library directories in the
generated Python wrappers.

Signed-off-by: Manolis Androulidakis <manolis@arrikto.com>
Reviewed-by: Ilias Tsitsimpis <iliastsi@arrikto.com>