Specify correct runtime library dependencies
Set the correct runtime library dependency for each library loaded on
Linux, based on the value of the loaded ELF's SONAME. Thus, generation
of wrappers can happen by loading the development libraries (.so) and
actual use of the wrappers can happen in a different environment, since
wrappers will try to load just the necessary ABI-compatible library
files (.so.X).

Note that in order to add the runtime library dependencies, we hijack
the 'options' during the processing phase (when finding the source library
for each exported function). This is done to avoid loading the same
library twice, which would slow the generation process.

Signed-off-by: Manolis Androulidakis <manolis@arrikto.com>
Reviewed-by: Ilias Tsitsimpis <iliastsi@arrikto.com>