Correctly handle dependencies for undef macros
Ctypsgen cannot handle the 'undef NULL' statement, present in the
stddef.h header file. The reason is that the NULL identifier is not
related with a MacroDescription as ctypesgen expects, because NULL is
related to the custom ConstantDescription that ctypesgen always defines.

Change the code that handles 'undef' directives to take into account
that the target identifier may not always be a MacroDescription but can
also be the NULL related to the custom-generated ConstantDescription.

After fixing the aforementioned bug, although ctypesgen may cancel any
definition of the NULL identifier in the Python module, when it meets an
'undef NULL' directive, a following definition has no effect. The reason
is that it does not remove the identifier from the dependencies lookup
table. This prevents new macro definitions of the identifier to be
included in the lookup table, so finally, any references to the
identifier after the undef macro, will be erroneously tied with the
original definition.

The NULL identifier is just an example. The same problem can occur for
any identifier defined in header1, undefined and redefined in header2
and referenced in a macro in header3, with header3 being the header file
for which ctypesgen generates wrappers.

Remove the associated identifier from the dependencies lookup table, so
that future re-definitions for that identifier have a chance to be added
to the dependencies, in case the identifier is referenced by any
following code. Essentially the identifier may be referenced only by any
macros that follow, since the C preprocessor has replaced the identifier
with its value for all C code. With this solution, the identifier
exported by the wrappers will be associated with the last macro
definition of that identifier, which is the desired outcome after all.

Signed-off-by: Manolis Androulidakis <manolis@arrikto.com>
Reviewed-by: Ilias Tsitsimpis <iliastsi@arrikto.com>