Enable in-memory module generation in unittests
Instead of writing the generated code into a .py file in
each test and then importing the module from file system,
the module is now directly generated from code in memory.
Therefore, the WrapperPrinter output has to be redirected to
a StringIO instance, as the WrapperPrinter does not support
writing the output to a variable. The redirected output is then
used to initialize a python module during runtime.