CommonHeaderTest has to be skipped on Windows
This test is currently not working on MS Windows. The reason is the call of
cleanup_common(), which tries to delete "common.dll" while it is still loaded
by ctypes.
"common.dll" gets loaded in multiple places: first, every call of
ctypesgen.processor.process() in generate_common() loads the dll to check
availability of the symbols mentioned in the header file.
In addition, each test imports the module generated by ctypesgen, which again
means the dll is loaded to memory.
On Linux/macOS this is no problem, as .so libaries can be overwritten/deleted
on file system while still loaded to memory.